I developed a Hangman game with multiple lengths of strings.
I developed a Hangman game with multiple lengths of strings. The game is simple and easy to use with the help of a mouse and a keyboard. My application has be able to keep scores of users and store them on a file. For usability purposes, I implemented a graphical user interface for my application. This project has been prepared using the java language.